requirements.txt
torch>=2.2
numpy>=1.23
pandas>=2.0
scikit-learn>=1.3
rdkit-pypi>=2022.9.5
transformers>=4.38
h5py>=3.10
matplotlib>=3.8
tqdm>=4.66
pyyaml>=6.0
